fixed polyline factor and sector_from clr
The "factor" attribute of the swoopy polyine indicates the degree to which the polyline will curve. I created the function getRanFloat() to yield a random float from 0.1 to 0.8*rand float and to check that the float does not already exist to ensure no line will overlap each other completely (as was the case in earlier versions). Put the sector_from coloring back how it was originally, so now each line is colored based on the sector from which the emissions come.